refactor: move pool-detail containers
