+++ packages/web/src/models/pool/pool-detail-model.ts
@@ -1,3 +1,57 @@
 import { PoolModel } from "./pool-model";
 
 export type PoolDetailModel = PoolModel;
+
+export const initialPool: PoolDetailModel = {
+  tokenA: {
+    chainId: "",
+    createdAt: "",
+    name: "",
+    address: "",
+    path: "",
+    decimals: 4,
+    symbol: "",
+    logoURI: "",
+    type: "native",
+    priceID: "",
+  },
+  tokenB: {
+    chainId: "",
+    createdAt: "",
+    name: "",
+    address: "",
+    path: "",
+    decimals: 4,
+    symbol: "",
+    logoURI: "",
+    type: "native",
+    priceID: "",
+  },
+  incentiveType: "INCENTIVIZED",
+  tvl: "0",
+  tvlChange: 0,
+  volume24h: 0,
+  rewards24hUsd: 0,
+  id: "",
+  apr: "0",
+  fee: "",
+  feeUsd24h: 0,
+  currentTick: 0,
+  price: 0,
+  tokenABalance: 0,
+  tokenBBalance: 0,
+  tickSpacing: 0,
+  rewardTokens: [],
+  totalApr: 0,
+  poolPath: "",
+  liquidity: "",
+  volumeChange24h: 0,
+  feeApr: "",
+  stakingApr: "",
+  allTimeVolumeUsd: "",
+  priceRatio: {
+    "7d": [],
+    "30d": [],
+    all: [],
+  },
+};
