+++ packages/web/src/views/pool/pool-detail/PoolDetail.tsx
@@ -2,9 +2,6 @@ import React, { useEffect, useMemo, useRef } from "react";
 
 import Footer from "@components/common/footer/Footer";
 import HeaderContainer from "@containers/header-container/HeaderContainer";
-import MyLiquidityContainer from "@containers/my-liquidity-container/MyLiquidityContainer";
-import PoolPairInformationContainer from "@containers/pool-pair-information-container/PoolPairInformationContainer";
-import StakingContainer from "@containers/staking-container/StakingContainer";
 import useCustomRouter from "@hooks/common/use-custom-router";
 import { usePositionData } from "@hooks/common/use-position-data";
 import useUrlParam from "@hooks/common/use-url-param";
@@ -12,6 +9,9 @@ import { useWallet } from "@hooks/wallet/use-wallet";
 import { useGetPoolDetailByPath } from "@query/pools";
 import { addressValidationCheck } from "@utils/validation-utils";
 
+import MyLiquidityContainer from "./containers/my-liquidity-container/MyLiquidityContainer";
+import PoolPairInformationContainer from "./containers/pool-pair-information-container/PoolPairInformationContainer";
+import StakingContainer from "./containers/staking-container/StakingContainer";
 import PoolLayout from "./PoolLayout";
 
 const PoolDetail: React.FC =() => {
